pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gsm"
version = "0.0.1"
description = "La programmazione per le persone"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name="la-comunita", email="lacomunita@dal.basso" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy",
    "pandas",
    "pymongo",
    "Flask",
    "flask-cors",
    "python-dotenv",
    "passlib"
]

[project.urls]
Homepage = "https://github.com/PMG-t/CMGS"

[tool.setuptools.packages]
find = { where = ["src"], include = ["gsm*"] }